+++ collapsed_lda/comparisons/reuter_sklearn.py
@@ -9,8 +9,7 @@ import click
 from sklearn.decomposition import LatentDirichletAllocation
 from sklearn.feature_extraction.text import CountVectorizer
 
-from collapsed_lda.comparisons.print_sklearn import *
-from collapsed_lda.utility import *
+from collapsed_lda.utility.utility import *
 
 
 @click.command()
