+++ collapsed_lda/utility/gensim.py
