+++ collapsed_lda/utility/sklearn.py
