+++ collapsed_lda/utility/utility.py
